refactor: relocate Response Section comp
